+++ packages/starter/templates/addons/tester/pure/src/app.spec.tsx
@@ -1,0 +1,10 @@
+import renderer from 'react-test-renderer';
+import { App } from './app';
+
+it('React app renders correctly', async () => {
+  const comp = renderer.create(<App />);
+
+  const tree = comp.toJSON();
+
+  expect(tree).toMatchSnapshot();
+});